+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Doctestable example for `model_predict` operation.
 - Doctestable examples to `operation/mapping.py`
 - load function in high_level API
+- Doctestable examples to `db` operations.
 ### Changed
 - `Edit on Github` button now hidden for plugins.